GUI/lipo_4s_mon/requirements.txt
pyserial==3.4
PyQt5==5.11.1
